small code adjustments, add test for utils
