+++ .gitignore
@@ -5,3 +5,4 @@ __pycache__/
 coverage.xml
 unpaywall_cache
 build/
+tests/test_cache
